requirements.txt
Flask==0.12.2
google-cloud-storage==1.3.1
gunicorn==19.7.1